Add KScrollView support Small refactor
