+++ kakao/src/main/kotlin/com/agoda/kakao/scroll/ScrollViewActions.kt
@@ -1,9 +1,8 @@
 @file:Suppress("unused")
 
-package com.agoda.kakao.list
+package com.agoda.kakao.scroll
 
 import android.view.View
-import android.widget.AdapterView
 import android.widget.ScrollView
 import androidx.test.espresso.UiController
 import androidx.test.espresso.ViewAction
@@ -60,31 +59,5 @@ interface ScrollViewActions : ScrollableActions, SwipeableActions {
                 }
             }
         })
-    }
-
-    /**
-     * Returns the size of ScrollView
-     *
-     * @return size of adapter
-     *
-     * @see ScrollView
-     * @see AdapterView
-     */
-    fun getSize(): Int {
-        var size = 0
-
-        view.perform(object : ViewAction {
-            override fun getDescription() = "Get AdapterView adapter size"
-
-            override fun getConstraints() = Matchers.allOf(ViewMatchers.isAssignableFrom(AdapterView::class.java), ViewMatchers.isDisplayed())
-
-            override fun perform(uiController: UiController?, view: View?) {
-                if (view is AdapterView<*>) {
-                    size = view.count
-                }
-            }
-        })
-
-        return size
     }
 }
